+++ src/core/src/main/java/org/apache/jmeter/visualizers/SamplingStatCalculator.java
@@ -199,8 +199,8 @@ public class SamplingStatCalculator {
             rtime = res.getTime();
             cmean = (long)calculator.getMean();
             cstdv = (long)calculator.getStandardDeviation();
-            cmedian = calculator.getMedian().longValue();
-            cpercent = calculator.getPercentPoint( 0.500 ).longValue();
+            cmedian = calculator.getMedian();
+            cpercent = calculator.getPercentPoint(0.500);
 // TODO cpercent is the same as cmedian here - why? and why pass it to "distributionLine"?
             rbool = res.isSuccessful();
         }
@@ -298,7 +298,7 @@ public class SamplingStatCalculator {
     }
 
     public Number getMin() {
-        if (calculator.getMin().longValue() < 0) {
+        if (calculator.getMin() < 0) {
             return 0L;
         }
         return calculator.getMin();
