+++ src/protocol/jdbc/src/main/java/org/apache/jmeter/protocol/jdbc/AbstractJDBCTestElement.java
@@ -485,7 +485,7 @@ public abstract class AbstractJDBCTestElement extends AbstractTestElement implem
                 throw new SQLException("Invalid data type: "+jdbcType, e);
             }
         }
-        return entry.intValue();
+        return entry;
     }
